feat: updating delete services

+++ app/Services/PostCategory/DeletePostCategory.php
@@ -7,10 +7,21 @@ use Illuminate\Support\Facades\DB;
 
 class DeletePostCategory
 {
-    public function delete(PostCategory $postCategory)
+    public function delete(int $id)
     {
+        \Log::info('DeletePostCategory service called', ['id' => $id]);
+
         try {
             DB::beginTransaction();
+
+            $postCategory = PostCategory::findOrFail($id);
+
+            if ($postCategory->posts()->exists()) {
+                \Log::info('Category has related posts, cannot delete');
+                throw new \Exception('This category cannot be deleted because it is associated with one or more posts.');
+            }
+
+            \Log::info('Category has no related posts, proceeding with deletion');
 
             $postCategory->delete();
 
@@ -19,7 +30,9 @@ class DeletePostCategory
             return $postCategory;
         } catch (\Exception $e) {
             DB::rollback();
+            \Log::error('Error deleting category', ['error' => $e->getMessage()]);
             return $e->getMessage();
         }
     }
+
 }
